requirements.txt
streamlit
pandas
numpy
joblib
requests
folium
streamlit-folium
networkx
osmnx
python-dotenv
streamlit-folium
scikit-learn
shapely

